Added triggers for bismark_active
Given the replication and complexity and chance for error
inherent in the creation of all these tables, use a script
to generate them.

Also traceroute is mildly special in that it actually will use
the generated id for something useful.

I note that django may still have trouble parsing traceroute_hops.